refactor: split store/mod.rs into metadata + search modules
