+++ src/store/mod.rs
@@ -11,11 +11,15 @@
 //! - `calls` - Call graph storage and queries
 //! - `types` - Type dependency storage and queries
 //! - `migrations` - Database schema migrations
+//! - `metadata` - Metadata get/set and version validation
+//! - `search` - FTS search, name search, RRF fusion
 
 mod calls;
 mod chunks;
+mod metadata;
 mod migrations;
 mod notes;
+mod search;
 mod types;
 
 /// Helper types and embedding conversion functions.
@@ -24,7 +28,6 @@ mod types;
 /// types from `cqs::store` instead of accessing `cqs::store::helpers` directly.
 pub(crate) mod helpers;
 
-use std::collections::HashMap;
 use std::path::Path;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::RwLock;
@@ -127,11 +130,6 @@ pub use types::TypeGraph;
 
 /// A type usage relationship from a chunk.
 pub use types::TypeUsage;
-
-// Internal use
-use helpers::ChunkRow;
-
-use crate::nl::normalize_for_fts;
 
 /// Defense-in-depth sanitization for FTS5 query strings.
 ///
@@ -464,485 +462,6 @@ impl Store {
         })
     }
 
-    /// Validates and optionally migrates the database schema version to match the current expected version.
-    ///
-    /// Queries the metadata table for the stored schema version and compares it against the current version. If the stored version is older, attempts to migrate the schema. Returns an error if the stored version is newer than the current version (indicating the database is incompatible), if the schema is corrupted, or if migration fails without a supported migration path.
-    ///
-    /// # Arguments
-    ///
-    /// `path` - The file path to the database, used for error reporting.
-    ///
-    /// # Returns
-    ///
-    /// Returns `Ok(())` if the schema version is valid and matches the current version, or if migration succeeds. Returns `Err(StoreError)` if the schema is newer than supported, corrupted, or migration fails.
-    ///
-    /// # Errors
-    ///
-    /// - `StoreError::SchemaNewerThanCq` - The stored schema version is newer than the current version.
-    /// - `StoreError::Corruption` - The stored schema version is not a valid integer.
-    /// - `StoreError::SchemaMismatch` - Schema migration is not supported for the version difference.
-    /// - Other `StoreError` variants from database access or migration failures.
-    fn check_schema_version(&self, path: &Path) -> Result<(), StoreError> {
-        let path_str = path.display().to_string();
-        self.rt.block_on(async {
-            let row: Option<(String,)> =
-                match sqlx::query_as("SELECT value FROM metadata WHERE key = 'schema_version'")
-                    .fetch_optional(&self.pool)
-                    .await
-                {
-                    Ok(r) => r,
-                    Err(sqlx::Error::Database(e)) if e.message().contains("no such table") => {
-                        return Ok(());
-                    }
-                    Err(e) => return Err(e.into()),
-                };
-
-            let version: i32 = match row {
-                Some((s,)) => s.parse().map_err(|e| {
-                    StoreError::Corruption(format!(
-                        "schema_version '{}' is not a valid integer: {}",
-                        s, e
-                    ))
-                })?,
-                // EH-22: Missing key is OK — init() hasn't been called yet on a fresh DB.
-                // After init(), schema_version is guaranteed present.
-                None => 0,
-            };
-
-            if version > CURRENT_SCHEMA_VERSION {
-                return Err(StoreError::SchemaNewerThanCq(version));
-            }
-            if version < CURRENT_SCHEMA_VERSION && version > 0 {
-                // Attempt migration instead of failing
-                match migrations::migrate(&self.pool, version, CURRENT_SCHEMA_VERSION).await {
-                    Ok(()) => {
-                        tracing::info!(
-                            path = %path_str,
-                            from = version,
-                            to = CURRENT_SCHEMA_VERSION,
-                            "Schema migrated successfully"
-                        );
-                    }
-                    Err(StoreError::MigrationNotSupported(from, to)) => {
-                        // No migration available, fall back to original error
-                        return Err(StoreError::SchemaMismatch(path_str, from, to));
-                    }
-                    Err(e) => return Err(e),
-                }
-            }
-            Ok(())
-        })
-    }
-
-    /// Validates that the stored model name and embedding dimensions match the expected values.
-    ///
-    /// This method checks the metadata table in the database to ensure compatibility between the current application and previously stored data. It verifies both the model name and the embedding vector dimensions.
-    ///
-    /// # Returns
-    ///
-    /// Returns `Ok(())` if validation passes or if the metadata table doesn't exist yet. Returns `Err(StoreError)` if the stored model name or dimensions don't match the expected values, or if a database error occurs.
-    ///
-    /// # Errors
-    ///
-    /// Returns `StoreError::ModelMismatch` if the stored model name differs from the expected `MODEL_NAME`.
-    ///
-    /// Returns `StoreError::DimensionMismatch` if the stored embedding dimensions differ from `EXPECTED_DIMENSIONS`.
-    ///
-    /// Returns other `StoreError` variants if database access fails.
-    fn check_model_version(&self) -> Result<(), StoreError> {
-        self.rt.block_on(async {
-            // Check model name
-            let row: Option<(String,)> =
-                match sqlx::query_as("SELECT value FROM metadata WHERE key = 'model_name'")
-                    .fetch_optional(&self.pool)
-                    .await
-                {
-                    Ok(r) => r,
-                    Err(sqlx::Error::Database(e)) if e.message().contains("no such table") => {
-                        return Ok(());
-                    }
-                    Err(e) => return Err(e.into()),
-                };
-
-            let stored_model = row.map(|(s,)| s).unwrap_or_default();
-
-            if !stored_model.is_empty() && stored_model != MODEL_NAME {
-                return Err(StoreError::ModelMismatch(
-                    stored_model,
-                    MODEL_NAME.to_string(),
-                ));
-            }
-
-            // Check embedding dimensions
-            let dim_row: Option<(String,)> =
-                sqlx::query_as("SELECT value FROM metadata WHERE key = 'dimensions'")
-                    .fetch_optional(&self.pool)
-                    .await?;
-
-            if let Some((dim_str,)) = dim_row {
-                if let Ok(stored_dim) = dim_str.parse::<u32>() {
-                    if stored_dim != EXPECTED_DIMENSIONS {
-                        return Err(StoreError::DimensionMismatch(
-                            stored_dim,
-                            EXPECTED_DIMENSIONS,
-                        ));
-                    }
-                } else {
-                    return Err(StoreError::Corruption(format!(
-                        "dimensions metadata '{}' is not a valid integer",
-                        dim_str
-                    )));
-                }
-            }
-
-            Ok(())
-        })
-    }
-
-    /// Checks if the stored CQL version in the metadata table matches the current application version.
-    ///
-    /// Retrieves the `cq_version` value from the metadata table and compares it against the current package version. If versions differ, logs an informational message. Errors during version retrieval are logged at debug level but do not propagate, allowing the application to continue.
-    ///
-    /// # Arguments
-    ///
-    /// `&self` - Reference to the store instance with access to the database pool and runtime.
-    ///
-    /// # Errors
-    ///
-    /// Errors are caught and logged but not propagated. Database query failures are logged at debug level.
-    fn check_cq_version(&self) {
-        if let Err(e) = self.rt.block_on(async {
-            let row: Option<(String,)> =
-                match sqlx::query_as("SELECT value FROM metadata WHERE key = 'cq_version'")
-                    .fetch_optional(&self.pool)
-                    .await
-                {
-                    Ok(row) => row,
-                    Err(e) => {
-                        tracing::debug!(error = %e, "Failed to read cq_version from metadata");
-                        return Ok::<_, StoreError>(());
-                    }
-                };
-
-            let stored_version = row.map(|(s,)| s).unwrap_or_default();
-            let current_version = env!("CARGO_PKG_VERSION");
-
-            if !stored_version.is_empty() && stored_version != current_version {
-                tracing::info!(
-                    "Index created by cqs v{}, running v{}",
-                    stored_version,
-                    current_version
-                );
-            }
-            Ok::<_, StoreError>(())
-        }) {
-            tracing::debug!(error = %e, "check_cq_version failed");
-        }
-    }
-
-    /// Search FTS5 index for keyword matches.
-    ///
-    /// # Search Method Overview
-    ///
-    /// The Store provides several search methods with different characteristics:
-    ///
-    /// - **`search_fts`**: Full-text keyword search using SQLite FTS5. Returns chunk IDs.
-    ///   Best for: Exact keyword matches, symbol lookup by name fragment.
-    ///
-    /// - **`search_by_name`**: Definition search by function/struct name. Uses FTS5 with
-    ///   heavy weighting on the name column. Returns full `SearchResult` with scores.
-    ///   Best for: "Where is X defined?" queries.
-    ///
-    /// - **`search_filtered`** (in search.rs): Semantic search with optional language/path
-    ///   filters. Can use RRF hybrid search combining semantic + FTS scores.
-    ///   Best for: Natural language queries like "retry with exponential backoff".
-    ///
-    /// - **`search_filtered_with_index`** (in search.rs): Like `search_filtered` but uses
-    ///   HNSW/CAGRA vector index for O(log n) candidate retrieval instead of brute force.
-    ///   Best for: Large indexes (>5k chunks) where brute force is slow.
-    pub fn search_fts(&self, query: &str, limit: usize) -> Result<Vec<String>, StoreError> {
-        let _span = tracing::info_span!("search_fts", limit).entered();
-        let normalized_query = sanitize_fts_query(&normalize_for_fts(query));
-        if normalized_query.is_empty() {
-            tracing::debug!(
-                original_query = %query,
-                "Query normalized to empty string, returning no FTS results"
-            );
-            return Ok(vec![]);
-        }
-
-        self.rt.block_on(async {
-            let rows: Vec<(String,)> = sqlx::query_as(
-                "SELECT id FROM chunks_fts WHERE chunks_fts MATCH ?1 ORDER BY bm25(chunks_fts) LIMIT ?2",
-            )
-            .bind(&normalized_query)
-            .bind(limit as i64)
-            .fetch_all(&self.pool)
-            .await?;
-
-            Ok(rows.into_iter().map(|(id,)| id).collect())
-        })
-    }
-
-    /// Search for chunks by name (definition search).
-    ///
-    /// Searches the FTS5 name column for exact or prefix matches.
-    /// Use this for "where is X defined?" queries instead of semantic search.
-    pub fn search_by_name(
-        &self,
-        name: &str,
-        limit: usize,
-    ) -> Result<Vec<SearchResult>, StoreError> {
-        let _span = tracing::info_span!("search_by_name", %name, limit).entered();
-        let limit = limit.min(100);
-        let normalized = sanitize_fts_query(&normalize_for_fts(name));
-        if normalized.is_empty() {
-            return Ok(vec![]);
-        }
-
-        // Pre-lowercase query once for score_name_match_pre_lower (PF-3)
-        let lower_name = name.to_lowercase();
-
-        // Search name column specifically using FTS5 column filter
-        // Use * for prefix matching (e.g., "parse" matches "parse_config")
-        debug_assert!(
-            !normalized.contains('"'),
-            "sanitized query must not contain double quotes"
-        );
-        if normalized.contains('"') {
-            return Ok(vec![]);
-        }
-        let fts_query = format!("name:\"{}\" OR name:\"{}\"*", normalized, normalized);
-
-        self.rt.block_on(async {
-            let rows: Vec<_> = sqlx::query(
-                "SELECT c.id, c.origin, c.language, c.chunk_type, c.name, c.signature, c.content, c.doc, c.line_start, c.line_end, c.parent_id, c.parent_type_name
-                 FROM chunks c
-                 JOIN chunks_fts f ON c.id = f.id
-                 WHERE chunks_fts MATCH ?1
-                 ORDER BY bm25(chunks_fts, 10.0, 1.0, 1.0, 1.0) -- Heavy weight on name column
-                 LIMIT ?2",
-            )
-            .bind(&fts_query)
-            .bind(limit as i64)
-            .fetch_all(&self.pool)
-            .await?;
-
-            let mut results = rows
-                .into_iter()
-                .map(|row| {
-                    let chunk = ChunkSummary::from(ChunkRow::from_row(&row));
-                    let name_lower = chunk.name.to_lowercase();
-                    let score = helpers::score_name_match_pre_lower(&name_lower, &lower_name);
-                    SearchResult { chunk, score }
-                })
-                .collect::<Vec<_>>();
-
-            // Re-sort by name-match score (FTS bm25 ordering may differ)
-            results.sort_by(|a, b| b.score.total_cmp(&a.score));
-
-            Ok(results)
-        })
-    }
-
-    /// Compute RRF (Reciprocal Rank Fusion) scores for combining two ranked lists.
-    ///
-    /// Pre-allocates the HashMap with capacity for both input lists (PERF-28).
-    /// Input size varies (limit*3 semantic + limit*3 FTS) but is always known upfront.
-    pub(crate) fn rrf_fuse(
-        semantic_ids: &[&str],
-        fts_ids: &[String],
-        limit: usize,
-    ) -> Vec<(String, f32)> {
-        // K=60 is the standard RRF constant from the original paper.
-        // Higher K reduces the impact of rank differences (smoother fusion).
-        const K: f32 = 60.0;
-
-        let mut scores: HashMap<&str, f32> =
-            HashMap::with_capacity(semantic_ids.len() + fts_ids.len());
-
-        for (rank, id) in semantic_ids.iter().enumerate() {
-            // RRF formula: 1 / (K + rank). The + 1.0 converts 0-indexed enumerate()
-            // to 1-indexed ranks (first result = rank 1, not rank 0).
-            let contribution = 1.0 / (K + rank as f32 + 1.0);
-            *scores.entry(id).or_insert(0.0) += contribution;
-        }
-
-        for (rank, id) in fts_ids.iter().enumerate() {
-            // Same conversion: enumerate's 0-index -> RRF's 1-indexed rank
-            let contribution = 1.0 / (K + rank as f32 + 1.0);
-            *scores.entry(id.as_str()).or_insert(0.0) += contribution;
-        }
-
-        let mut sorted: Vec<(String, f32)> = scores
-            .into_iter()
-            .map(|(k, v)| (k.to_string(), v))
-            .collect();
-        sorted.sort_by(|a, b| b.1.total_cmp(&a.1));
-        sorted.truncate(limit);
-        sorted
-    }
-
-    /// Exposed for property testing only
-    #[cfg(test)]
-    pub(crate) fn rrf_fuse_test(
-        semantic_ids: &[String],
-        fts_ids: &[String],
-        limit: usize,
-    ) -> Vec<(String, f32)> {
-        let refs: Vec<&str> = semantic_ids.iter().map(|s| s.as_str()).collect();
-        Self::rrf_fuse(&refs, fts_ids, limit)
-    }
-
-    /// Update the `updated_at` metadata timestamp to now.
-    ///
-    /// Call after indexing operations complete (pipeline, watch reindex, note sync)
-    /// to track when the index was last modified.
-    pub fn touch_updated_at(&self) -> Result<(), StoreError> {
-        let now = chrono::Utc::now().to_rfc3339();
-        self.rt.block_on(async {
-            sqlx::query("INSERT OR REPLACE INTO metadata (key, value) VALUES ('updated_at', ?1)")
-                .bind(&now)
-                .execute(&self.pool)
-                .await?;
-            Ok(())
-        })
-    }
-
-    /// Mark the HNSW index as dirty (out of sync with SQLite).
-    ///
-    /// Call before writing chunks to SQLite. Clear after successful HNSW save.
-    /// On load, a dirty flag means a crash occurred between SQLite commit and
-    /// HNSW save — the HNSW index should not be trusted.
-    pub fn set_hnsw_dirty(&self, dirty: bool) -> Result<(), StoreError> {
-        let val = if dirty { "1" } else { "0" };
-        self.rt.block_on(async {
-            sqlx::query("INSERT OR REPLACE INTO metadata (key, value) VALUES ('hnsw_dirty', ?1)")
-                .bind(val)
-                .execute(&self.pool)
-                .await?;
-            Ok(())
-        })
-    }
-
-    /// Check if the HNSW index is marked as dirty (potentially stale).
-    ///
-    /// Returns `false` if the key doesn't exist (pre-v13 indexes).
-    pub fn is_hnsw_dirty(&self) -> Result<bool, StoreError> {
-        self.rt.block_on(async {
-            let row: Option<(String,)> =
-                sqlx::query_as("SELECT value FROM metadata WHERE key = 'hnsw_dirty'")
-                    .fetch_optional(&self.pool)
-                    .await?;
-            Ok(row.is_some_and(|(v,)| v == "1"))
-        })
-    }
-
-    /// Set a metadata key/value pair, or delete it if `value` is `None`.
-    fn set_metadata_opt(&self, key: &str, value: Option<&str>) -> Result<(), StoreError> {
-        self.rt.block_on(async {
-            match value {
-                Some(v) => {
-                    sqlx::query("INSERT OR REPLACE INTO metadata (key, value) VALUES (?1, ?2)")
-                        .bind(key)
-                        .bind(v)
-                        .execute(&self.pool)
-                        .await?;
-                }
-                None => {
-                    sqlx::query("DELETE FROM metadata WHERE key = ?1")
-                        .bind(key)
-                        .execute(&self.pool)
-                        .await?;
-                }
-            }
-            Ok(())
-        })
-    }
-
-    /// Get a metadata value by key, returning `None` if the key doesn't exist.
-    fn get_metadata_opt(&self, key: &str) -> Result<Option<String>, StoreError> {
-        self.rt.block_on(async {
-            let row: Option<(String,)> =
-                sqlx::query_as("SELECT value FROM metadata WHERE key = ?1")
-                    .bind(key)
-                    .fetch_optional(&self.pool)
-                    .await?;
-            Ok(row.map(|(v,)| v))
-        })
-    }
-
-    /// Store a pending LLM batch ID so interrupted processes can resume polling.
-    pub fn set_pending_batch_id(&self, batch_id: Option<&str>) -> Result<(), StoreError> {
-        self.set_metadata_opt("pending_llm_batch", batch_id)
-    }
-
-    /// Get the pending LLM batch ID, if any.
-    pub fn get_pending_batch_id(&self) -> Result<Option<String>, StoreError> {
-        self.get_metadata_opt("pending_llm_batch")
-    }
-
-    /// Store a pending doc-comment batch ID so interrupted processes can resume polling.
-    pub fn set_pending_doc_batch_id(&self, batch_id: Option<&str>) -> Result<(), StoreError> {
-        self.set_metadata_opt("pending_doc_batch", batch_id)
-    }
-
-    /// Get the pending doc-comment batch ID, if any.
-    pub fn get_pending_doc_batch_id(&self) -> Result<Option<String>, StoreError> {
-        self.get_metadata_opt("pending_doc_batch")
-    }
-
-    /// Store a pending HyDE batch ID so interrupted processes can resume polling.
-    pub fn set_pending_hyde_batch_id(&self, batch_id: Option<&str>) -> Result<(), StoreError> {
-        self.set_metadata_opt("pending_hyde_batch", batch_id)
-    }
-
-    /// Get the pending HyDE batch ID, if any.
-    pub fn get_pending_hyde_batch_id(&self) -> Result<Option<String>, StoreError> {
-        self.get_metadata_opt("pending_hyde_batch")
-    }
-
-    /// Get cached notes summaries (loaded on first call, invalidated on mutation).
-    ///
-    /// Returns a cloned Vec rather than a slice reference to avoid holding the
-    /// RwLock read guard across caller code. The clone cost is negligible — notes
-    /// are typically <100 entries with small strings.
-    pub fn cached_notes_summaries(&self) -> Result<Vec<NoteSummary>, StoreError> {
-        {
-            let guard = self.notes_summaries_cache.read().unwrap_or_else(|p| {
-                tracing::warn!("notes cache read lock poisoned, recovering");
-                p.into_inner()
-            });
-            if let Some(ref ns) = *guard {
-                return Ok(ns.clone());
-            }
-        }
-        // Cache miss — load from DB and populate
-        let ns = self.list_notes_summaries()?;
-        {
-            let mut guard = self.notes_summaries_cache.write().unwrap_or_else(|p| {
-                tracing::warn!("notes cache write lock poisoned, recovering");
-                p.into_inner()
-            });
-            *guard = Some(ns.clone());
-        }
-        Ok(ns)
-    }
-
-    /// Invalidate the cached notes summaries.
-    ///
-    /// Must be called after any operation that modifies notes (upsert, replace, delete)
-    /// so subsequent reads see fresh data.
-    pub(crate) fn invalidate_notes_cache(&self) {
-        match self.notes_summaries_cache.write() {
-            Ok(mut guard) => *guard = None,
-            Err(p) => {
-                tracing::warn!("notes cache write lock poisoned during invalidation, recovering");
-                *p.into_inner() = None;
-            }
-        }
-    }
-
     /// Gracefully close the store, performing WAL checkpoint.
     ///
     /// This ensures all WAL changes are written to the main database file,
@@ -1005,111 +524,11 @@ mod tests {
     use super::*;
     use proptest::prelude::*;
 
-    // ===== Property-based tests for RRF =====
+    use crate::nl::normalize_for_fts;
+
+    // ===== FTS fuzz tests =====
 
     proptest! {
-        /// Property: RRF scores are always positive
-        #[test]
-        fn prop_rrf_scores_positive(
-            semantic in prop::collection::vec("[a-z]{1,5}", 0..20),
-            fts in prop::collection::vec("[a-z]{1,5}", 0..20),
-            limit in 1usize..50
-        ) {
-            let result = Store::rrf_fuse_test(&semantic, &fts, limit);
-            for (_, score) in &result {
-                prop_assert!(*score > 0.0, "RRF score should be positive: {}", score);
-            }
-        }
-
-        /// Property: RRF scores are bounded
-        /// Note: Duplicates in input lists can accumulate extra points.
-        /// Max theoretical: sum of 1/(K+r+1) for all appearances across both lists.
-        #[test]
-        fn prop_rrf_scores_bounded(
-            semantic in prop::collection::vec("[a-z]{1,5}", 0..20),
-            fts in prop::collection::vec("[a-z]{1,5}", 0..20),
-            limit in 1usize..50
-        ) {
-            let result = Store::rrf_fuse_test(&semantic, &fts, limit);
-            // Conservative upper bound: sum of first N terms of 1/(K+r+1) for both lists
-            // where N is max list length (20). With duplicates, actual max is ~0.3
-            let max_possible = 0.5; // generous bound accounting for duplicates
-            for (id, score) in &result {
-                prop_assert!(
-                    *score <= max_possible,
-                    "RRF score {} for '{}' exceeds max {}",
-                    score, id, max_possible
-                );
-            }
-        }
-
-        /// Property: RRF respects limit
-        #[test]
-        fn prop_rrf_respects_limit(
-            semantic in prop::collection::vec("[a-z]{1,5}", 0..30),
-            fts in prop::collection::vec("[a-z]{1,5}", 0..30),
-            limit in 1usize..20
-        ) {
-            let result = Store::rrf_fuse_test(&semantic, &fts, limit);
-            prop_assert!(
-                result.len() <= limit,
-                "Result length {} exceeds limit {}",
-                result.len(), limit
-            );
-        }
-
-        /// Property: RRF results are sorted by score descending
-        #[test]
-        fn prop_rrf_sorted_descending(
-            semantic in prop::collection::vec("[a-z]{1,5}", 1..20),
-            fts in prop::collection::vec("[a-z]{1,5}", 1..20),
-            limit in 1usize..50
-        ) {
-            let result = Store::rrf_fuse_test(&semantic, &fts, limit);
-            for window in result.windows(2) {
-                prop_assert!(
-                    window[0].1 >= window[1].1,
-                    "Results not sorted: {} < {}",
-                    window[0].1, window[1].1
-                );
-            }
-        }
-
-        /// Property: Items appearing in both lists get higher scores
-        /// Note: Uses hash_set to ensure unique IDs - duplicates in input lists
-        /// accumulate scores which can violate the "overlap wins" property.
-        #[test]
-        fn prop_rrf_rewards_overlap(
-            common_id in "[a-z]{3}",
-            only_semantic in prop::collection::hash_set("[A-Z]{3}", 1..5),
-            only_fts in prop::collection::hash_set("[0-9]{3}", 1..5)
-        ) {
-            let mut semantic = vec![common_id.clone()];
-            semantic.extend(only_semantic);
-            let mut fts = vec![common_id.clone()];
-            fts.extend(only_fts);
-
-            let result = Store::rrf_fuse_test(&semantic, &fts, 100);
-
-            let common_score = result.iter()
-                .find(|(id, _)| id == &common_id)
-                .map(|(_, s)| *s)
-                .unwrap_or(0.0);
-
-            let max_single = result.iter()
-                .filter(|(id, _)| id != &common_id)
-                .map(|(_, s)| *s)
-                .fold(0.0f32, |a, b| a.max(b));
-
-            prop_assert!(
-                common_score >= max_single,
-                "Common item score {} should be >= single-list max {}",
-                common_score, max_single
-            );
-        }
-
-        // ===== FTS fuzz tests =====
-
         #[test]
         fn fuzz_normalize_for_fts_no_panic(input in "\\PC{0,500}") {
             let _ = normalize_for_fts(&input);
@@ -1260,119 +679,7 @@ mod tests {
         }
     }
 
-    // ===== TC-8: pending batch ID =====
-
-    use crate::test_helpers::setup_store;
-
-    #[test]
-    fn test_pending_batch_roundtrip() {
-        let (store, _dir) = setup_store();
-        store.set_pending_batch_id(Some("batch_123")).unwrap();
-        let result = store.get_pending_batch_id().unwrap();
-        assert_eq!(result, Some("batch_123".to_string()));
-    }
-
-    #[test]
-    fn test_pending_batch_clear() {
-        let (store, _dir) = setup_store();
-        store.set_pending_batch_id(Some("batch_abc")).unwrap();
-        store.set_pending_batch_id(None).unwrap();
-        let result = store.get_pending_batch_id().unwrap();
-        assert_eq!(result, None);
-    }
-
-    #[test]
-    fn test_pending_batch_default_none() {
-        let (store, _dir) = setup_store();
-        let result = store.get_pending_batch_id().unwrap();
-        assert_eq!(result, None);
-    }
-
-    #[test]
-    fn test_pending_batch_overwrite() {
-        let (store, _dir) = setup_store();
-        store.set_pending_batch_id(Some("a")).unwrap();
-        store.set_pending_batch_id(Some("b")).unwrap();
-        let result = store.get_pending_batch_id().unwrap();
-        assert_eq!(result, Some("b".to_string()));
-    }
-
-    // ===== TC-10: HNSW dirty flag =====
-
-    #[test]
-    fn test_hnsw_dirty_roundtrip() {
-        let (store, _dir) = setup_store();
-        store.set_hnsw_dirty(true).unwrap();
-        assert!(store.is_hnsw_dirty().unwrap());
-    }
-
-    #[test]
-    fn test_hnsw_dirty_default_false() {
-        let (store, _dir) = setup_store();
-        assert!(!store.is_hnsw_dirty().unwrap());
-    }
-
-    #[test]
-    fn test_hnsw_dirty_toggle() {
-        let (store, _dir) = setup_store();
-        store.set_hnsw_dirty(true).unwrap();
-        assert!(store.is_hnsw_dirty().unwrap());
-
-        store.set_hnsw_dirty(false).unwrap();
-        assert!(!store.is_hnsw_dirty().unwrap());
-
-        store.set_hnsw_dirty(true).unwrap();
-        assert!(store.is_hnsw_dirty().unwrap());
-    }
-
-    // ===== TC-16: cache invalidation =====
-
-    #[test]
-    fn test_cached_notes_empty() {
-        let (store, _dir) = setup_store();
-        let notes = store.cached_notes_summaries().unwrap();
-        assert!(notes.is_empty());
-    }
-
-    #[test]
-    fn test_cached_notes_invalidation() {
-        let (store, dir) = setup_store();
-
-        let source = dir.path().join("notes.toml");
-        std::fs::write(&source, "# dummy").unwrap();
-
-        // Insert first batch of notes
-        let note1 = crate::note::Note {
-            id: "note:0".to_string(),
-            text: "first note".to_string(),
-            sentiment: 0.0,
-            mentions: vec![],
-        };
-        store.upsert_notes_batch(&[note1], &source, 100).unwrap();
-
-        // Populate cache
-        let cached = store.cached_notes_summaries().unwrap();
-        assert_eq!(cached.len(), 1);
-        assert_eq!(cached[0].text, "first note");
-
-        // Replace notes with a different set (replace_notes_for_file invalidates cache)
-        let note2 = crate::note::Note {
-            id: "note:0".to_string(),
-            text: "replaced note".to_string(),
-            sentiment: 0.5,
-            mentions: vec!["src/lib.rs".to_string()],
-        };
-        store
-            .replace_notes_for_file(&[note2], &source, 200)
-            .unwrap();
-
-        // Cache should reflect the replacement
-        let cached = store.cached_notes_summaries().unwrap();
-        assert_eq!(cached.len(), 1);
-        assert_eq!(cached[0].text, "replaced note");
-    }
-
-    // ===== TC-17: check_model_version tests =====
+    // ===== TC-19: concurrent access and edge-case tests =====
 
     fn make_test_store_initialized() -> (Store, tempfile::TempDir) {
         let dir = tempfile::TempDir::new().unwrap();
@@ -1381,112 +688,6 @@ mod tests {
         store.init(&ModelInfo::default()).unwrap();
         (store, dir)
     }
-
-    #[test]
-    fn tc17_model_mismatch_returns_error() {
-        let (store, _dir) = make_test_store_initialized();
-        store
-            .set_metadata_opt("model_name", Some("wrong-model/v1"))
-            .unwrap();
-        let err = store.check_model_version().unwrap_err();
-        assert!(
-            matches!(err, StoreError::ModelMismatch(..)),
-            "Expected ModelMismatch, got: {:?}",
-            err
-        );
-    }
-
-    #[test]
-    fn tc17_dimension_mismatch_returns_error() {
-        let (store, _dir) = make_test_store_initialized();
-        store.set_metadata_opt("dimensions", Some("999")).unwrap();
-        let err = store.check_model_version().unwrap_err();
-        assert!(
-            matches!(err, StoreError::DimensionMismatch(999, _)),
-            "Expected DimensionMismatch, got: {:?}",
-            err
-        );
-    }
-
-    #[test]
-    fn tc17_corrupt_dimension_returns_corruption() {
-        let (store, _dir) = make_test_store_initialized();
-        store
-            .set_metadata_opt("dimensions", Some("not_a_number"))
-            .unwrap();
-        let err = store.check_model_version().unwrap_err();
-        assert!(
-            matches!(err, StoreError::Corruption(..)),
-            "Expected Corruption, got: {:?}",
-            err
-        );
-    }
-
-    #[test]
-    fn tc17_correct_model_passes() {
-        let (store, _dir) = make_test_store_initialized();
-        assert!(store.check_model_version().is_ok());
-    }
-
-    // ===== TC-18: check_schema_version tests =====
-
-    #[test]
-    fn tc18_schema_newer_returns_error() {
-        let (store, _dir) = make_test_store_initialized();
-        let future_version = (CURRENT_SCHEMA_VERSION + 1).to_string();
-        store
-            .set_metadata_opt("schema_version", Some(&future_version))
-            .unwrap();
-        let err = store
-            .check_schema_version(std::path::Path::new("/test"))
-            .unwrap_err();
-        assert!(
-            matches!(err, StoreError::SchemaNewerThanCq(..)),
-            "Expected SchemaNewerThanCq, got: {:?}",
-            err
-        );
-    }
-
-    #[test]
-    fn tc18_corrupt_schema_version_returns_corruption() {
-        let (store, _dir) = make_test_store_initialized();
-        store
-            .set_metadata_opt("schema_version", Some("garbage"))
-            .unwrap();
-        let err = store
-            .check_schema_version(std::path::Path::new("/test"))
-            .unwrap_err();
-        assert!(
-            matches!(err, StoreError::Corruption(..)),
-            "Expected Corruption, got: {:?}",
-            err
-        );
-    }
-
-    #[test]
-    fn tc18_current_schema_passes() {
-        let (store, _dir) = make_test_store_initialized();
-        assert!(store
-            .check_schema_version(std::path::Path::new("/test"))
-            .is_ok());
-    }
-
-    #[test]
-    fn tc18_missing_schema_key_passes() {
-        // Fresh DB with metadata table but no schema_version key
-        let (store, _dir) = make_test_store_initialized();
-        store.rt.block_on(async {
-            sqlx::query("DELETE FROM metadata WHERE key = 'schema_version'")
-                .execute(&store.pool)
-                .await
-                .unwrap();
-        });
-        assert!(store
-            .check_schema_version(std::path::Path::new("/test"))
-            .is_ok());
-    }
-
-    // ===== TC-19: concurrent access and edge-case tests =====
 
     #[test]
     fn concurrent_readonly_opens() {
